% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilisticMod.R
\name{probabilisticMod}
\alias{probabilisticMod}
\title{probabilisticMod function}
\usage{
probabilisticMod(data, opt, param)
}
\arguments{
\item{data}{Input data, given as a matrix.  Must contain unbroken, aligned columns of a) dates, b) observed streamflow, and c) simulated streamflow from a deterministic model.}

\item{opt}{List of options.}

\item{param}{List of parameters.}

\item{opt$reps}{Number of probabilistic replicates to be generated. Recommended (and default) number is 100. Higher numbers of replicates require more computing time and lower numbers risk innacurate predictions.}

\item{opt$dirname}{Directory into which the output files are to be saved.}

\item{opt$title}{Default 'replicate'. The title (minus the extension) of the printed .pdf that is output from this function.}

\item{opt$obs}{Column name of the observed streamflow in the input data. Default is 'obs'. This input must match the column header of the observed streamflow.}

\item{opt$pred}{Column name of the simulated, deterministic streamflow in the input data. Default is 'pred'. This input must match the column header of the simulated streamflow.}

\item{opt$date}{Column name of the dates in the input data. Default is 'date'. This input must match the column header of the dates.}

\item{opt$meantype}{Options are "zero" or "linear", with "linear" being default. Defines the structure of the mean parameter in the probabilistic model. "zero" uses traditional modelling assumptions, and "linear" represents an innovative new approach that is more generally-applicable to a wider range of objective functions. Please refer to Hunter et al. 2021 for a comprehensive demonstration of the difference in predictive quality between the two.}

\item{opt$unit}{Units of the input streamflow. Default is 'mmd'. Alternatives are 'mm/d', 'm3s', 'm3/s','ML/d', 'MLd'.}

\item{opt$repPrint}{T / F to print a .csv containing the probabilistic replicates. Default is 'T'.}

\item{opt$plPrint}{T / F to print a .csv containing the probability limits. Default is 'T'.}

\item{param$A}{Box-Cox shift parameter.}

\item{param$lambda}{Box-Cox power parameter.}
}
\description{
Generates error predictions from a given observed and simulated series. Outputs a .pdf summary file containing plots and metrics, and (optionally) two .csv files containing the probabilistic predictions themselves and the probability limits. By J Hunter & Team.
}
\examples{
probabilisticMod(data,opt,param)             ## An example of how to launch the function.

data = read.csv('inputDataFile.csv',as.is=T) ## Example of setting up the data file.

opt = list(reps=100,
           title='myProbPredictions',
           obs='observedData',
           pred='predictedData',
           date='dates',
           meantype='linear',
           unit='ML/d',
           repPrint=T,
           plPrint=T)                        ## Example of setting up the user options.

param = list(A=0.,
             lambda=0.2)                     ## Example of setting up the input parameters.
}
\keyword{none}
