% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probModInteractive.R
\name{probModInteractive}
\alias{probModInteractive}
\title{probModInteractive function}
\usage{
probModInteractive()
}
\description{
Loads the probabilistic model to an interactive R-Shiny interface. By J Hunter & Team.
}
\examples{
probModInteractive()          ## An example of how to use the function. All user inputs are accessed from within the Shiny interface.
}
\keyword{none}
