% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probPredInteractive.R
\name{probPredInteractive}
\alias{probPredInteractive}
\title{probPredInteractive function}
\usage{
probPredInteractive()
}
\description{
Loads the probabilistic model to an interactive R-Shiny interface. By J Hunter & Team.
}
\examples{
probPredInteractive()          ## An example of how to use the function. All user inputs are accessed from within the Shiny interface.
}
\keyword{none}
