% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_message.R
\name{xerr}
\alias{xerr}
\title{Error message repository}
\usage{
xerr(flag)
}
\arguments{
\item{flag}{# indicates which error is being called}
}
\description{
For internal use with the shiny app only. By J Hunter & Team.
}
\examples{
none          ## no example - this script is not for the user; it is for the webapp - and webapps need no examples
}
\keyword{none}
