% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilisticMod.R
\name{probabilisticMod}
\alias{probabilisticMod}
\title{probabilisticMod function}
\usage{
probabilisticMod(data, opt)
}
\arguments{
\item{data}{# input data, given as a .csv file.  Must contain unbroken, aligned columns of a) dates, b) observed streamflow, and c) simulated streamflow.}

\item{opt}{# list of options.  All options must be given for now.}
}
\description{
Generates error predictions from a given observed and simulated series. By J Hunter & Team.
}
\examples{
probabilisticMod(data,opt)          ## An example of how to use the function
}
\keyword{none}
